+++ core/src/main/java/io/substrait/function/ToTypeString.java
@@ -82,12 +82,12 @@ public class ToTypeString
 
   @Override
   public String visit(final Type.IntervalYear expr) {
-    return "year";
+    return "iyear";
   }
 
   @Override
   public String visit(final Type.IntervalDay expr) {
-    return "day";
+    return "iday";
   }
 
   @Override
@@ -113,6 +113,16 @@ public class ToTypeString
   @Override
   public String visit(final Type.Decimal expr) {
     return "dec";
+  }
+
+  @Override
+  public String visit(final Type.PrecisionTimestamp expr) {
+    return "pts";
+  }
+
+  @Override
+  public String visit(final Type.PrecisionTimestampTZ expr) {
+    return "ptstz";
   }
 
   @Override
@@ -153,6 +163,16 @@ public class ToTypeString
   @Override
   public String visit(ParameterizedType.Decimal expr) throws RuntimeException {
     return "dec";
+  }
+
+  @Override
+  public String visit(ParameterizedType.PrecisionTimestamp expr) throws RuntimeException {
+    return "pts";
+  }
+
+  @Override
+  public String visit(ParameterizedType.PrecisionTimestampTZ expr) throws RuntimeException {
+    return "ptstz";
   }
 
   @Override