+++ core/src/main/java/io/substrait/type/TypeVisitor.java
@@ -23,9 +23,15 @@ public interface TypeVisitor<R, E extends Throwable> {
 
   R visit(Type.Time type) throws E;
 
+  @Deprecated
   R visit(Type.TimestampTZ type) throws E;
 
+  @Deprecated
   R visit(Type.Timestamp type) throws E;
+
+  R visit(Type.PrecisionTimestamp type) throws E;
+
+  R visit(Type.PrecisionTimestampTZ type) throws E;
 
   R visit(Type.IntervalYear type) throws E;
 
@@ -159,6 +165,16 @@ public interface TypeVisitor<R, E extends Throwable> {
 
     @Override
     public R visit(Type.Decimal type) throws E {
+      throw t();
+    }
+
+    @Override
+    public R visit(Type.PrecisionTimestamp type) throws E {
+      throw t();
+    }
+
+    @Override
+    public R visit(Type.PrecisionTimestampTZ type) throws E {
       throw t();
     }