+++ core/src/test/java/io/substrait/type/proto/AggregateRoundtripTest.java
@@ -6,6 +6,7 @@ import io.substrait.TestBase;
 import io.substrait.expression.AggregateFunctionInvocation;
 import io.substrait.expression.Expression;
 import io.substrait.expression.ExpressionCreator;
+import io.substrait.expression.FunctionOption;
 import io.substrait.expression.ImmutableExpression;
 import io.substrait.extension.ExtensionCollector;
 import io.substrait.relation.Aggregate;
@@ -13,6 +14,7 @@ import io.substrait.relation.ImmutableAggregate;
 import io.substrait.relation.ProtoRelConverter;
 import io.substrait.relation.RelProtoConverter;
 import io.substrait.relation.VirtualTableScan;
+import io.substrait.type.NamedStruct;
 import io.substrait.type.TypeCreator;
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -25,8 +27,12 @@ public class AggregateRoundtripTest extends TestBase {
   private void assertAggregateRoundtrip(Expression.AggregationInvocation invocation) {
     var expression = ExpressionCreator.decimal(false, BigDecimal.TEN, 10, 2);
     Expression.StructLiteral literal =
-        ImmutableExpression.StructLiteral.builder().from(expression).build();
-    var input = VirtualTableScan.builder().addRows(literal).build();
+        ImmutableExpression.StructLiteral.builder().addFields(expression).build();
+    var input =
+        VirtualTableScan.builder()
+            .initialSchema(NamedStruct.of(Arrays.asList("decimal"), R.struct(R.decimal(10, 2))))
+            .addRows(literal)
+            .build();
     ExtensionCollector functionCollector = new ExtensionCollector();
     var to = new RelProtoConverter(functionCollector);
     var extensions = defaultExtensionCollection;
@@ -41,6 +47,12 @@ public class AggregateRoundtripTest extends TestBase {
                     .outputType(TypeCreator.of(false).I64)
                     .aggregationPhase(Expression.AggregationPhase.INITIAL_TO_RESULT)
                     .invocation(invocation)
+                    .options(
+                        Arrays.asList(
+                            FunctionOption.builder()
+                                .name("option")
+                                .addValues("VALUE1", "VALUE2")
+                                .build()))
                     .build())
             .build();