+++ isthmus/src/main/java/io/substrait/isthmus/expression/LiteralConverter.java
@@ -104,8 +104,8 @@ public class LiteralConverter {
         }
         throw new UnsupportedOperationException("Unable to handle char type: " + val);
       }
-      case DOUBLE -> fp64(n, bd(literal).doubleValue());
-      case FLOAT -> fp32(n, bd(literal).floatValue());
+      case FLOAT, DOUBLE -> fp64(n, bd(literal).doubleValue());
+      case REAL -> fp32(n, bd(literal).floatValue());
 
       case DECIMAL -> {
         BigDecimal bd = bd(literal);